Added rocrand_generate_uniform

+++ crush_test/file_common.hpp
@@ -34,9 +34,9 @@ void rocrand_file_write_results(std::string file_name, T * data, size_t n)
 		for(int i = 0; i < n; i++)
 		{   
             if (i == n - 1)
-                fout << data[i];
+                fout << static_cast<float>(data[i]);
             else
-                fout << data[i] << " ";
+                fout << static_cast<float>(data[i]) << " ";
 		}
         std::cout << "File was written successfully" << std::endl;
 	}
